requirements.txt
numpy>=1.21.0
cupy-cuda12x>=12.0.0
numba>=0.57.0
matplotlib>=3.5.0
scipy>=1.9.0
